Chore(batching): handle exceptions out of parade dispatcher
